fix distribution and move folders
