+++ src/components/Contact/ContactElements/Form.js
@@ -15,56 +15,56 @@ const StyledTextField = styled(TextField)`
     &.Mui-focused fieldset {
       border-color: rgb(216, 216, 216);} }`;
 const useStyles = makeStyles(({  
-    General:{
+    general:{
         display: "flex",
         flexFlow: "row wrap"},
-    FormControl:{
+    formControl:{
         width: "100%",
         marginBottom: 20,},
-    TextField:{
+    textField:{
         margin: "0px 20px 0px 20px"},
-    Button:{
+    button:{
         marginLeft: "20px"},
-    TextArea:{
+    textArea:{
         width: "100%",
         margin: "0px 20px 20px 20px"},
-    Label:{
+    label:{
         margin: "0px 0px 12px 20px"}
         }));
 function Form(){
 const classes = useStyles();
 return (
     <div>
-        <form className={classes.General}>                         
+        <form className={classes.general}>                         
             <Grid item xs={12} sm={6} md={6}>
-                <FormControl className={classes.FormControl}>
-                    <Typography className={classes.Label} variant="body2">Nombre</Typography>
-                    <StyledTextField className={classes.TextField} variant="outlined" size="small"/>
+                <FormControl className={classes.formControl}>
+                    <Typography className={classes.label} variant="body2">Nombre</Typography>
+                    <StyledTextField className={classes.textField} variant="outlined" size="small"/>
                 </FormControl>
             </Grid>                
             <Grid item xs={12} sm={6} md={6}>
-                <FormControl className={classes.FormControl}>
-                    <Typography className={classes.Label} variant="body2">Email</Typography>                     
-                    <StyledTextField className={classes.TextField} variant="outlined" size="small"/>                    
+                <FormControl className={classes.formControl}>
+                    <Typography className={classes.label} variant="body2">Email</Typography>                     
+                    <StyledTextField className={classes.textField} variant="outlined" size="small"/>                    
                 </FormControl>
             </Grid>
             <Grid item xs={12} sm={6} md={6}>
-                <FormControl className={classes.FormControl}>
-                    <Typography className={classes.Label}  variant="body2">Telefono</Typography>                
-                    <StyledTextField className={classes.TextField} variant="outlined" size="small"/>                 
+                <FormControl className={classes.formControl}>
+                    <Typography className={classes.label}  variant="body2">Telefono</Typography>                
+                    <StyledTextField className={classes.textField} variant="outlined" size="small"/>                 
                 </FormControl>
             </Grid>
             <Grid item xs={12} sm={6} md={6}>
-                <FormControl className={classes.FormControl}>
-                    <Typography className={classes.Label} variant="body2">Asunto</Typography>                 
-                    <StyledTextField className={classes.TextField} variant="outlined" size="small"/>         
+                <FormControl className={classes.formControl}>
+                    <Typography className={classes.label} variant="body2">Asunto</Typography>                 
+                    <StyledTextField className={classes.textField} variant="outlined" size="small"/>         
                 </FormControl>
             </Grid>                                                
-                <FormControl className={classes.TextArea}>
+                <FormControl className={classes.textArea}>
                     <Box mb={1.5}><Typography variant="body2">Mensaje</Typography></Box>
                     <StyledTextField variant="outlined" multiline rows={6}/>
                 </FormControl>                   
-            <Button className={classes.Button} variant="contained">Enviar</Button>
+            <Button className={classes.button} variant="contained">Enviar</Button>
         </form>
     </div>
     )
