+++ src/components/Contact/ContactElements/TitleBox.js
@@ -5,21 +5,24 @@ import Typography from '@material-ui/core/Typography';
 import Divider from '@material-ui/core/Divider';
 
 const useStyles = makeStyles(({
-    Typography:{
+    typography:{
             fontSize: 25,
             fontFamily: "Dancing Script, cursive",
             display: 'flex',
-            justifyContent: 'center'},
-    TitleBox:{
+            justifyContent: 'center',
+            color: "#ad172b" 
+        },
+    titleBox:{
             width: "70%",
-            margin: "0% 0% 7% 15%"},
+            margin: "0% 0% 7% 15%"
+        },
     }));
 function TitleBox(props){
     const classes = useStyles();
     return (
         <div>
-            <Box className={classes.TitleBox}>  
-                <Typography className={classes.Typography} variant="h6">{props.children}</Typography><Divider/> 
+            <Box className={classes.titleBox}>  
+                <Typography className={classes.typography} variant="h6">{props.children}</Typography><Divider/> 
             </Box>
         </div>
     )
