+++ src/components/NavegationBar/NavegationBar.js
@@ -5,7 +5,7 @@ import Divider from '@material-ui/core/Divider'
 import AppBar from '@material-ui/core/AppBar';
 import Button from '@material-ui/core/Button';
 import Drawer from '@material-ui/core/Drawer';
-import logo from '../assets/logo.png'
+import logo from '../../assets/logo.png'
 import ShoppingCartIcon from '@material-ui/icons/ShoppingCart';
 import MenuIcon from '@material-ui/icons/Menu';
 import Hidden from '@material-ui/core/Hidden';
@@ -47,9 +47,11 @@ const useStyles = makeStyles(theme => ({
         margin: '0 10px'
     },
     logo: {
-        width: 150,
+        width: 75,
+        margin: 5,
+        borderRadius: "100%",
         ['@media (max-width:500px)']: {
-            width: 95,
+            width: 65,
         },
     },
     contentShopping: {
