+++ src/router/index.js
@@ -1,11 +1,12 @@
 import React from 'react';
 import { BrowserRouter, Switch, Route, Link } from "react-router-dom";
-import NavegationBar from '../components/NavegationBar'
-import Home from '../components/Home'
-import { makeStyles } from '@material-ui/core/styles';
-import Contact from '../components/Contact';
+import NavegationBar from '../components/NavegationBar/NavegationBar';
+import Home from '../components/Home';
+import  makeStyles  from '@material-ui/core/styles/makeStyles';
+import Contact from '../components/Contact/Contact';
+import Footer from '../components/Footer/Footer' 
 
- const useStyles = makeStyles(theme => ({
+ /*const useStyles = makeStyles(theme => ({
     root: {
       flexGrow: 1,
       background: '#263238',
@@ -15,7 +16,7 @@ import Contact from '../components/Contact';
       color: '#fff',
       textDecoration: 'unset',
     },
- }));
+ }));*/
 
 export default function Router() {
   return (
@@ -25,6 +26,7 @@ export default function Router() {
             <Route exact path={'/'} component={Home}/>
             <Route exact path='/Contact' component={Contact}/>
         </Switch>
+        <Footer/>
     </BrowserRouter>
   );
 }
